Add a module parameter for specifying the table type.
Can also be used by utilities such as dsrtools to know what table
the module has been configured for by reading the file
/sys/module/xt_DADDR/parameters/table.

Note that yvipagent was not adjusted to account for the change of
the default table.  The tool is expected to be deprecated shortly
in favor of dsrtools.