Remove pre and preun checks for kernel module being unused.
When I added this check feature back 10 years ago, it was to prevent
problems from user errors where they'd upgrade the package, but not
immediately see the new behaviors because in between stopping and
restarting their L3DSR software rules, they failed to unload the old
module.  Nowdays though this check is a hindrance for doing fully
automated, in-place production upgrades.