Remove the check to skip changes when the new address is the same as the old address.
The check should never be false in normal use (new should never
equal old), so it is an unncessary optimization that slows down the
common path.  Also, without the check, it simplifies testing of the
algorithms where we can take advantage making new == old when we
don't have a load balancer to use.